requirements.txt
streamlit
pillow
pytesseract
easyocr
opencv-python-headless
numpy
requests
